Add floating-point scaling
Also adds a new cvar called scr_scale: -1 for old behavior,
-2 for fractional, and anything from .1 to 20

+++ src/y_inter.c
@@ -537,7 +537,7 @@ void Y_IntermissionDrawer(void)
 
 		if (animatetic && (tic_t)intertic >= animatetic)
 		{
-			const INT32 scradjust = (vid.width/vid.dup)>>3; // 40 for BASEVIDWIDTH
+			const INT32 scradjust = (INT32)(vid.width/vid.dup)>>3; // 40 for BASEVIDWIDTH
 			INT32 animatetimer = (intertic - animatetic);
 			if (animatetimer <= 16)
 			{
